Refactored RoomService
