+++ hotel-booking-app-service/src/main/java/com/application/hotelbooking/services/RoomCreationService.java
@@ -1,0 +1,7 @@
+package com.application.hotelbooking.services;
+
+import com.application.hotelbooking.dto.RoomCreationServiceDTO;
+
+public interface RoomCreationService {
+    void createRoomFromDTO(RoomCreationServiceDTO roomCreationServiceDTO);
+}
