+++ hotel-booking-app-service/src/main/java/com/application/hotelbooking/services/RoomService.java
@@ -8,5 +8,4 @@ import java.util.List;
 
 public interface RoomService {
     List<ReservableRoomDTO> searchRooms(RoomSearchFormServiceDTO roomSearchFormServiceDTO);
-    void createRoomFromDTO(RoomCreationServiceDTO roomCreationServiceDTO);
 }
